+++ src/routes/extractors/lavabo_extractors.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub(crate) struct CreateLavaboRequest {
     pub(crate) descripcio: String,
-    pub(crate) puntuacio: f32,
+    pub(crate) puntuacio: f64,
 }
